removed CMWorkspace
