+++ sdk/cloudmachine/Azure.Provisioning.CloudMachine/src/Tooling/Azd.cs
@@ -32,7 +32,7 @@ public static class Azd
         Directory.CreateDirectory(infraDirectory);
 
         infra.Build().Save(infraDirectory);
-        var cmid = CloudMachineWorkspace.ReadOrCreateCloudMachineId();
+        var cmid = CloudMachineClient.ReadOrCreateCloudMachineId();
 
         // main.bicep
         var location = new ProvisioningParameter("location", typeof(string));
